+++ e2e/coverage-handlebars/__tests__/greet.js
@@ -8,6 +8,6 @@ const greet = require('../greet.hbs');
 
 test('am', () => {
   expect(greet({am: true, name: 'Joe'})).toEqual(
-    '<p>Good\n  morning\nJoe!</p>\n'
+    '<p>Good\n  morning\nJoe!</p>\n',
   );
 });