+++ packages/jest-runtime/src/index.ts
@@ -19,7 +19,7 @@ import {
   compileFunction,
 } from 'vm';
 import * as nativeModule from 'module';
-import type {Config} from '@jest/types';
+import type {Config, Global} from '@jest/types';
 import type {
   Jest,
   JestEnvironment,
@@ -27,6 +27,7 @@ import type {
   Module,
   ModuleWrapper,
 } from '@jest/environment';
+import type * as JestGlobals from '@jest/globals';
 import type {SourceMapRegistry} from '@jest/source-map';
 import {formatStackTrace, separateMessageFromStack} from 'jest-message-util';
 import {createDirectory, deepCyclicCopy} from 'jest-util';
@@ -51,6 +52,11 @@ import HasteMap = require('jest-haste-map');
 import Resolver = require('jest-resolve');
 import Snapshot = require('jest-snapshot');
 import stripBOM = require('strip-bom');
+
+interface JestGlobalsValues extends Global.TestFrameworkGlobals {
+  jest: JestGlobals.jest;
+  expect: JestGlobals.expect;
+}
 
 type HasteMapOptions = {
   console?: Console;
@@ -133,7 +139,8 @@ class Runtime {
   private _moduleMocker: typeof jestMock;
   private _isolatedModuleRegistry: ModuleRegistry | null;
   private _moduleRegistry: ModuleRegistry;
-  private _esmoduleRegistry: Map<string, VMModule>;
+  private _esmoduleRegistry: Map<string, Promise<VMModule>>;
+  private _needsCoverageMapped: Set<string>;
   private _resolver: Resolver;
   private _shouldAutoMock: boolean;
   private _shouldMockModuleCache: BooleanObject;
@@ -147,6 +154,8 @@ class Runtime {
   private _unmockList: RegExp | undefined;
   private _virtualMocks: BooleanObject;
   private _moduleImplementation?: typeof nativeModule.Module;
+  private jestObjectCaches: Map<string, Jest>;
+  private _hasWarnedAboutRequireCacheModification = false;
 
   constructor(
     config: Config.ProjectConfig,
@@ -183,6 +192,7 @@ class Runtime {
     this._sourceMapRegistry = Object.create(null);
     this._fileTransforms = new Map();
     this._virtualMocks = Object.create(null);
+    this.jestObjectCaches = new Map();
 
     this._mockMetaDataCache = Object.create(null);
     this._shouldMockModuleCache = Object.create(null);
@@ -315,13 +325,6 @@ class Runtime {
     modulePath: Config.Path,
     query = '',
   ): Promise<VMModule> {
-    if (modulePath === '@jest/globals') {
-      // TODO: create a Synthetic Module for this. Will need to create a `jest` object without a `LocalModuleRequire`
-      throw new Error(
-        'Importing `@jest/globals` is not supported from ESM yet',
-      );
-    }
-
     const cacheKey = modulePath + query;
 
     if (!this._esmoduleRegistry.has(cacheKey)) {
@@ -355,11 +358,16 @@ class Runtime {
         },
       });
 
-      this._esmoduleRegistry.set(cacheKey, module);
-
-      await module.link(this.linkModules.bind(this));
-
-      await module.evaluate();
+      this._esmoduleRegistry.set(
+        cacheKey,
+        // we wanna put the linking promise in the cache so modules loaded in
+        // parallel can all await it. We then await it synchronously below, so
+        // we shouldn't get any unhandled rejections
+        module
+          .link(this.linkModules.bind(this))
+          .then(() => module.evaluate())
+          .then(() => module),
+      );
     }
 
     const module = this._esmoduleRegistry.get(cacheKey);
@@ -369,11 +377,27 @@ class Runtime {
     return module;
   }
 
-  private async linkModules(specifier: string, referencingModule: VMModule) {
+  private linkModules(specifier: string, referencingModule: VMModule) {
+    if (specifier === '@jest/globals') {
+      const fromCache = this._esmoduleRegistry.get('@jest/globals');
+
+      if (fromCache) {
+        return fromCache;
+      }
+      const globals = this.getGlobalsForEsm(
+        referencingModule.identifier,
+        referencingModule.context,
+      );
+      this._esmoduleRegistry.set('@jest/globals', globals);
+
+      return globals;
+    }
+
     const resolved = this._resolveModule(
       referencingModule.identifier,
       specifier,
     );
+
     if (
       this._resolver.isCoreModule(resolved) ||
       this.unstable_shouldLoadAsEsm(resolved)
@@ -647,12 +671,18 @@ class Runtime {
     };
   }
 
-  requireModuleOrMock(from: Config.Path, moduleName: string): unknown {
+  requireModuleOrMock<T = unknown>(from: Config.Path, moduleName: string): T {
+    // this module is unmockable
+    if (moduleName === '@jest/globals') {
+      // @ts-ignore: we don't care that it's not assignable to T
+      return this.getGlobalsForCjs(from);
+    }
+
     try {
       if (this._shouldMock(from, moduleName)) {
-        return this.requireMock(from, moduleName);
+        return this.requireMock<T>(from, moduleName);
       } else {
-        return this.requireModule(from, moduleName);
+        return this.requireModule<T>(from, moduleName);
       }
     } catch (e) {
       const moduleNotFound = Resolver.tryCastModuleNotFoundError(e);
@@ -968,6 +998,10 @@ class Runtime {
       return;
     }
 
+    const jestObject = this._createJestObjectFor(filename);
+
+    this.jestObjectCaches.set(filename, jestObject);
+
     try {
       compiledFunction.call(
         localModule.exports,
@@ -977,10 +1011,7 @@ class Runtime {
         dirname, // __dirname
         filename, // __filename
         this._environment.global, // global object
-        this._createJestObjectFor(
-          filename,
-          localModule.require as LocalModuleRequire,
-        ), // jest object
+        jestObject, // jest object
         ...this._config.extraGlobals.map(globalVariable => {
           if (this._environment.global[globalVariable]) {
             return this._environment.global[globalVariable];
@@ -1257,7 +1288,13 @@ class Runtime {
     moduleRequire.resolve = resolve;
     moduleRequire.cache = (() => {
       const notPermittedMethod = () => {
-        console.warn('`require.cache` modification is not permitted');
+        if (!this._hasWarnedAboutRequireCacheModification) {
+          this._environment.global.console.warn(
+            '`require.cache` modification is not permitted',
+          );
+
+          this._hasWarnedAboutRequireCacheModification = true;
+        }
         return true;
       };
       return new Proxy<RequireCache>(Object.create(null), {
@@ -1295,10 +1332,7 @@ class Runtime {
     return moduleRequire;
   }
 
-  private _createJestObjectFor(
-    from: Config.Path,
-    localRequire: LocalModuleRequire,
-  ): Jest {
+  private _createJestObjectFor(from: Config.Path): Jest {
     const disableAutomock = () => {
       this._shouldAutoMock = false;
       return jestObject;
@@ -1429,8 +1463,8 @@ class Runtime {
       isMockFunction: this._moduleMocker.isMockFunction,
       isolateModules,
       mock,
-      requireActual: localRequire.requireActual,
-      requireMock: localRequire.requireMock,
+      requireActual: this.requireActual.bind(this, from),
+      requireMock: this.requireMock.bind(this, from),
       resetAllMocks,
       resetModuleRegistry: resetModules,
       resetModules,
@@ -1509,6 +1543,69 @@ class Runtime {
     }
 
     throw e;
+  }
+
+  private getGlobalsForCjs(from: Config.Path): JestGlobalsValues {
+    const jest = this.jestObjectCaches.get(from);
+
+    invariant(jest, 'There should always be a Jest object already');
+
+    return {...this.getGlobalsFromEnvironment(), jest};
+  }
+
+  private async getGlobalsForEsm(
+    from: Config.Path,
+    context: VMContext,
+  ): Promise<VMModule> {
+    let jest = this.jestObjectCaches.get(from);
+
+    if (!jest) {
+      jest = this._createJestObjectFor(from);
+
+      this.jestObjectCaches.set(from, jest);
+    }
+
+    const globals: JestGlobalsValues = {
+      ...this.getGlobalsFromEnvironment(),
+      jest,
+    };
+
+    const module = new SyntheticModule(
+      Object.keys(globals),
+      function () {
+        Object.entries(globals).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      {context, identifier: '@jest/globals'},
+    );
+
+    await module.link(() => {
+      throw new Error('This should never happen');
+    });
+
+    await module.evaluate();
+
+    return module;
+  }
+
+  private getGlobalsFromEnvironment(): Omit<JestGlobalsValues, 'jest'> {
+    return {
+      afterAll: this._environment.global.afterAll,
+      afterEach: this._environment.global.afterEach,
+      beforeAll: this._environment.global.beforeAll,
+      beforeEach: this._environment.global.beforeEach,
+      describe: this._environment.global.describe,
+      expect: this._environment.global.expect,
+      fdescribe: this._environment.global.fdescribe,
+      fit: this._environment.global.fit,
+      it: this._environment.global.it,
+      test: this._environment.global.test,
+      xdescribe: this._environment.global.xdescribe,
+      xit: this._environment.global.xit,
+      xtest: this._environment.global.xtest,
+    };
   }
 }